FEATURE: exports all contents and renames folder from models to types
