+++ libs/spirit-islander/util/src/index.ts
@@ -1,1 +1,24 @@
-export * from './lib/spirit-islander-util';
+// Data
+export * from './lib/data/adversaries';
+export * from './lib/data/boards';
+export * from './lib/data/expansions';
+export * from './lib/data/maps';
+export * from './lib/data/scenarios';
+export * from './lib/data/spirits';
+
+// App types
+export * from './lib/types/combo.interface';
+export * from './lib/types/config.interface';
+export * from './lib/types/game-setup.interface';
+export * from './lib/types/page.enum';
+
+// Game types
+export * from './lib/types/game/adversaries';
+export * from './lib/types/game/board';
+export * from './lib/types/game/difficulty';
+export * from './lib/types/game/expansions';
+export * from './lib/types/game/maps';
+export * from './lib/types/game/option';
+export * from './lib/types/game/players';
+export * from './lib/types/game/scenarios';
+export * from './lib/types/game/spirits';
