+++ libs/spirit-islander/util/src/lib/data/boards.ts
@@ -1,4 +1,4 @@
-import type { Board } from '../models/game/board';
+import type { Board } from '../types/game/board';
 
 export const BOARDS: Board[] = [
   {
