--- libs/spirit-islander/util/src/lib/models/spirit-islander-util.spec.ts
@@ -1,7 +1,0 @@
-import { spiritIslanderUtil } from './spirit-islander-util';
-
-describe('spiritIslanderUtil', () => {
-  it('should work', () => {
-    expect(spiritIslanderUtil()).toEqual('spirit-islander-util');
-  });
-});
