--- libs/spirit-islander/util/src/lib/models/spirit-islander-util.ts
@@ -1,3 +1,0 @@
-export function spiritIslanderUtil(): string {
-  return 'spirit-islander-util';
-}
