+++ libs/spirit-islander/util/src/lib/types/config.interface.ts
