+++ libs/spirit-islander/util/src/lib/types/game-setup.interface.ts
