+++ libs/spirit-islander/util/src/lib/types/game/adversaries.ts
