+++ libs/spirit-islander/util/src/lib/types/game/board.ts
