+++ libs/spirit-islander/util/src/lib/types/game/difficulty.ts
