+++ libs/spirit-islander/util/src/lib/types/game/maps.ts
