+++ libs/spirit-islander/util/src/lib/types/game/option.ts
