+++ libs/spirit-islander/util/src/lib/types/game/scenarios.ts
