+++ libs/spirit-islander/util/src/lib/types/game/spirits.ts
